+++ ATSAccessibility/Navigators/WaterNavigator.cs
@@ -124,11 +124,7 @@ namespace ATSAccessibility.Navigators {
 
 		protected override bool PerformSubItemAction(int sectionIndex, int itemIndex, int subItemIndex) {
 			if (_sectionTypes[sectionIndex] == SectionType.Workers) {
-				if (_workersSection.PerformSubItemAction(itemIndex, subItemIndex)) {
-					_navigationLevel = 1;
-					return true;
-				}
-				return false;
+				return PerformWorkerSubItemAction(itemIndex, subItemIndex);
 			}
 
 			if (_sectionTypes[sectionIndex] == SectionType.Upgrades) {